# Added Chinese plugin information, changed style of checking allowed chars