--- a/src/main/resources/config.yml
+++ b/src/main/resources/config.yml
@@ -174,7 +174,7 @@ guild:
  startslots: 10
  #distance from spawn
  fromspawn: 100
- allowedchars: ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789
+ allowedchars: "[A-Za-z0-9]*"
  settings:
   tag:
    #colored tags