config.yml
StringUtils.java
CommandNovaGuilds.java