@MarcinWieczorek 2016-01-02
Thanks for contribution. 
I'm gonna review your changes. 
Did you check the efficiency of regular expressions and my old code? I'm going to test that first.