@MarcinWieczorek 2016-01-02
Performance tests:
```---------------------------------------------------------
Generating strings...
Generated 10000000 strings in 6.144s
---------------------------------------------------------
Testing Old method
Old method finished in 0.553s
---------------------------------------------------------
Testing New method
New method finished in 4.692s
---------------------------------------------------------
Difference:
4.139s
OLD method is faster
---------------------------------------------------------
```