@berry64 2016-01-02
The old method may be faster, but when i tried to use chinese/special unicode characters i cant, so by using the match way, i think i can use more characters