@berry64 2016-01-02
does the old method support special characters?
eg. chinese characters