@MarcinWieczorek 2016-01-02
Well, how would it be possible to match Chinese characters with the 'normal' alphabet?
Detecting disallowed characters wouldn't make sense, there's always a character you didn't list, there are too many... And I guess listing allowed Chinese characters is not possible as well.
Ideas?