@berry64 2016-01-02
maybe add a switch to config allowing users to use any characters?
so instead of checking, it returns true directly perhaps