@MarcinWieczorek 2016-01-02
Seems legit.
```YAML
guild:
 stringchecking:
  enabled: true
  regex: false
  pattern: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
  regexpattern: "[A-Za-z0-9]*"
```

Let's connect both methods and add an 'off switch' like you said.
Wanna do it?
PS: change stringchecking to whatever, had no better idea.