@MarcinWieczorek 2016-01-02
Also, please make another pull request with your /ng translation, and I'm closing this one.