@berry64 2016-01-02
to the both methods: sure
to another pull request: done, created another one